+++ backend/pottery/libs/database/src/migrations/1757433007377-CreateStoreTable.ts
@@ -3,30 +3,38 @@ import { MigrationInterface, QueryRunner, Table } from "typeorm";
 export class CreateStoreTable1757433007377 implements MigrationInterface {
 
     public async up(queryRunner: QueryRunner): Promise<void> {
-        await queryRunner.createTable(
-            new Table({
-                name: 'stores',
-                columns: [
-                    {
-                        name: 'id',
-                        type: 'int',
-                        isPrimary: true,
-                        isGenerated: true,
-                        generationStrategy: 'increment',
-                    },
-                    { name: 'store_name', type: 'varchar', isNullable: false, isUnique: true },
-                    { name: 'address', type: 'varchar', isNullable: true },
-                    { name: 'phone', type: 'varchar', length: '12', isNullable: true },
-                    { name: 'created_at', type: 'timestamptz', default: 'CURRENT_TIMESTAMP' },
-                    { name: 'updated_at', type: 'timestamptz', isNullable: true, default: 'CURRENT_TIMESTAMP' },
-                    { name: 'deleted_at', type: 'timestamptz', isNullable: true },
-                ],
-            }),
-        )
+        const tableExists = await queryRunner.hasTable('stores');
+        if (!tableExists) {
+            await queryRunner.createTable(
+                new Table({
+                    name: 'stores',
+                    columns: [
+                        {
+                            name: 'id',
+                            type: 'int',
+                            isPrimary: true,
+                            isGenerated: true,
+                            generationStrategy: 'increment',
+                        },
+                        { name: 'store_name', type: 'varchar', isNullable: false, isUnique: true },
+                        { name: 'address', type: 'varchar', isNullable: true },
+                        { name: 'phone', type: 'varchar', length: '12', isNullable: true },
+                        { name: 'created_at', type: 'timestamptz', default: 'CURRENT_TIMESTAMP' },
+                        { name: 'updated_at', type: 'timestamptz', isNullable: true, default: 'CURRENT_TIMESTAMP' },
+                        { name: 'deleted_at', type: 'timestamptz', isNullable: true },
+                    ],
+                }),
+            );
+        } else {
+            console.log('⚠️  Table "stores" already exists — skipping creation.');
+        }
     }
 
     public async down(queryRunner: QueryRunner): Promise<void> {
-        await queryRunner.dropTable('stores')
+        const tableExists = await queryRunner.hasTable('stores');
+        if (tableExists) {
+            await queryRunner.dropTable('stores');
+        }
     }
 
 }
